requirements.txt
azure-core==1.21.1
azure-storage-queue==12.1.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cycler==0.11.0
fonttools==4.31.2
greenlet==1.1.2
idna==3.3
isodate==0.6.1
kiwisolver==1.4.2
matplotlib==3.5.1
msrest==0.6.21
networkx==2.7.1
numpy==1.22.2
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
